Allow for Vector2Int to be divided using a float

+++ Grid-Astar-Library/code/Vector2Extensions.cs
@@ -9,7 +9,7 @@ public static partial class Vector2Extensions
 	/// <param name="self"></param>
 	/// <param name="divideBy"></param>
 	/// <returns></returns>
-	public static IntVector2 ToIntVector2( this Vector2 self, int divideBy = 1 )
+	public static IntVector2 ToIntVector2( this Vector2 self, float divideBy = 1 )
 	{
 
 		return new IntVector2( (int)Math.Round( self.x / divideBy ), (int)Math.Round( self.y / divideBy ) );
@@ -22,7 +22,7 @@ public static partial class Vector2Extensions
 	/// <param name="self"></param>
 	/// <param name="divideBy"></param>
 	/// <returns></returns>
-	public static IntVector2 ToIntVector2( this Vector3 self, int divideBy = 1 )
+	public static IntVector2 ToIntVector2( this Vector3 self, float divideBy = 1f )
 	{
 
 		return new IntVector2( (int)Math.Round( self.x / divideBy ), (int)Math.Round( self.y / divideBy ) );
